Raw model finished
I did the first model to test the cryptographic module, than I made the second script (attempt) that worked with the flaws. The third script I created can locally store a key, an encrypted username, and an encrypted password and use the key to decrypt it. Once you login and change your username and password, script wil close it-self. So you need to re-open it and than changes are made.